rework logic to use dictionary based item spawning, selling and transport and have a central building manager that fires commands rather than the items themselves
buildings now inherrit from base building class with setup method
buildings now use building manager for certain logic
items no longer call logic but recive orders instead

+++ Assets/Scripts/Managers/UiManager.cs
@@ -6,7 +6,7 @@ public class UiManager : SingletonBase<UiManager>
 {
    public TMP_Text currentCurrency;
 
-   public void UpdateCurrency(int value)
+   public void UpdateCurrency(float value)
    {
       currentCurrency.text = value.ToString();
    }
